Update 08-44_testicular-and-paratesticular-tumors-in-children.md
In the section “Pathogenesis and Overview of Testicular and Paratesticular Tumor Types in Children” Paragraph 9: It must be remembered that the prepubertal testis is a common location for metastatic or infiltrative malignancy, such as metastatic leukaemia and lymphoma. It should read : It must be remembered that the testis in childhood is a common location for metastatic or infiltrative malignancy, such as metastatic leukemia or lymphoma. In Table 1: Last Row, Paratesticular Tumours, Under Mostly Benign, May be malignant: Melanotic neuroectodermal tumor of Infancy is one entity (It reads as three entities because of the dividing lines)